+++ src/hr/avrbanac/openglplayground/renderers/MasterRenderer.java
@@ -24,7 +24,7 @@ import org.lwjgl.opengl.GL11;
  * optimization will be used while rendering terrain.
  * 
  * @author avrbanac
- * @version 1.0.4
+ * @version 1.0.10
  */
 public class MasterRenderer {
     private StaticShader shader;
@@ -60,19 +60,19 @@ public class MasterRenderer {
         GL11.glDisable(GL11.GL_CULL_FACE);
     }
     
-    public void render(Light sun, Camera camera) {
+    public void render(List<Light> lights, Camera camera) {
         prepare();
         
         shader.start();
         shader.loadSkyColor(SKY_RED, SKY_GREEN, SKY_BLUE);
-        shader.loadLight(sun);
+        shader.loadLights(lights);
         shader.loadViewMatrix(camera);
         renderer.render(entities);
         shader.stop();
         
         terrainShader.start();
         terrainShader.loadSkyColor(SKY_RED, SKY_GREEN, SKY_BLUE);
-        terrainShader.loadLight(sun);
+        terrainShader.loadLights(lights);
         terrainShader.loadViewMatrix(camera);
         terrainRenderer.render(terrains);
         terrainShader.stop();