+++ src/hr/avrbanac/openglplayground/shaders/StaticShader.java
@@ -7,20 +7,21 @@ import hr.avrbanac.openglplayground.entities.Light;
 import hr.avrbanac.openglplayground.maths.Matrix4f;
 import hr.avrbanac.openglplayground.maths.Vector2f;
 import hr.avrbanac.openglplayground.maths.Vector3f;
+import java.util.List;
 
 /**
  * This is just an implementation of abstract class {@ShaderProgram}
  * 
  * @author avrbanac
- * @version 1.0.8
+ * @version 1.0.10
  */
 public class StaticShader extends ShaderProgram {
 
     private int locationTransformationMatrix;
     private int locationProjectionMatrix;
     private int locationViewMatrix;
-    private int locationLightPosition;
-    private int locationLightColor;
+    private int locationLightPosition[];
+    private int locationLightColor[];
     private int locationShineDamper;
     private int locationReflectivity;
     private int locationViewMatrixInverse;
@@ -45,8 +46,8 @@ public class StaticShader extends ShaderProgram {
         locationTransformationMatrix    = super.getUniformLocation("transformationMatrix");
         locationProjectionMatrix        = super.getUniformLocation("projectionMatrix");
         locationViewMatrix              = super.getUniformLocation("viewMatrix");
-        locationLightPosition           = super.getUniformLocation("lightPosition");
-        locationLightColor              = super.getUniformLocation("lightColor");
+//        locationLightPosition           = super.getUniformLocation("lightPosition");
+//        locationLightColor              = super.getUniformLocation("lightColor");
         locationShineDamper             = super.getUniformLocation("shineDamper");
         locationReflectivity            = super.getUniformLocation("reflectivity");
         locationUseFakeLighting         = super.getUniformLocation("useFakeLighting");
@@ -57,6 +58,13 @@ public class StaticShader extends ShaderProgram {
         //this can be done in vertex shader using inverse of viewMatrix if GLSL version supports it
         locationViewMatrixInverse       = super.getUniformLocation("viewMatrixInv");
         
+        // this part changed since we have arrays now instead of one light source
+        locationLightPosition           = new int[MAX_LIGHTS_NUMBER];
+        locationLightColor              = new int[MAX_LIGHTS_NUMBER];
+        for(int i = 0; i < MAX_LIGHTS_NUMBER; i++) {
+            locationLightPosition[i] = super.getUniformLocation("lightPosition[" + i + "]");
+            locationLightColor[i] = super.getUniformLocation("lightColor[" + i + "]");
+        }
     }
     
     public void loadTransformationMatrix(Matrix4f matrix) {
@@ -75,9 +83,17 @@ public class StaticShader extends ShaderProgram {
     }
     
     // diffusal lighting
-    public void loadLight(Light light) {
-        super.load3DVector(locationLightPosition, light.getPosition());
-        super.load3DVector(locationLightColor, light.getColor());
+    public void loadLights(List<Light> lights) {
+        for (int i = 0; i < MAX_LIGHTS_NUMBER; i++) {
+            if(i<lights.size()) {
+                super.load3DVector(locationLightPosition[i], lights.get(i).getPosition());
+                super.load3DVector(locationLightColor[i], lights.get(i).getColor());
+            } else {
+                // if there is not enough lights in list load up empty info
+                super.load3DVector(locationLightPosition[i], new Vector3f(0,0,0));
+                super.load3DVector(locationLightColor[i], new Vector3f(0,0,0));
+            }
+        }
     }
     
     // specular lighting